refactor(frontend): extract fetch-in-useEffect into custom hooks
Move data fetching logic out of useEffect into useConversation and
useConversations hooks with proper AbortController support. Eliminates
react-doctor fetch-in-useEffect errors.
